+++ lib/interactor/contracts/terms.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-require 'dry-validation'
 require 'interactor/contracts/outcome'
 
 module Interactor
@@ -13,8 +12,8 @@ module Interactor
       #   terms = Interactor::Contracts::Terms.new
       #
       # @api public
-      # @param [Dry::Validation::Schema] terms the terms to start with
-      def initialize(terms = Class.new(Dry::Validation::Schema))
+      # @param [Dry::Validation::Contract] terms the terms to start with
+      def initialize(terms = Class.new(Dry::Validation::Contract))
         @terms = terms
       end
 
@@ -30,7 +29,9 @@ module Interactor
       # @param [Block] term the term to add to the terms
       # @return [void]
       def add(&term)
-        @terms = Dry::Validation.Schema(@terms, { build: false }, &term)
+        @terms = Class.new(Dry::Validation::Contract).tap do |new_terms|
+          new_terms.params(@terms.schema, &term)
+        end
       end
 
       # Validates the terms against a given context
@@ -46,7 +47,20 @@ module Interactor
       # @param [#to_h] context the context to validate the terms against
       # @return [Outcome]
       def call(context)
+        define_empty_schema
+
         Outcome.new(@terms.new.call(context.to_h))
+      end
+
+      private
+
+      # Defines no-op rules block if no schema has been added
+      # Prevents the raising of Dry::Validation::SchemaMissingError exception
+      #
+      # @api private
+      # @return [void]
+      def define_empty_schema
+        @terms.params.nil? && @terms.params {}
       end
     end
   end